Add files via upload
This is a dashboard, a simple web application designed to display data analysis results. It references material and concepts from a Dicoding course, making use of essential skills in data visualization and interactive web development. The goal is to present insights clearly and concisely, helping users understand and interpret data trends effectively.